requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
langchain-google-genai
langgraph
sqlalchemy>=2.0.23
asyncpg>=0.29.0
greenlet>=3.0.0
pydantic>=2.5.2
python-dotenv>=1.0.0
langchain-experimental
wikipedia
